# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ARCC Signal Hub operational API
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorResponse {
  error: string;
}

export type HealthStatusTelegram = typeof HealthStatusTelegram[keyof typeof HealthStatusTelegram];


export const HealthStatusTelegram = {
  connected: 'connected',
  disconnected: 'disconnected',
  not_configured: 'not_configured',
} as const;

export type HealthStatusDatabase = typeof HealthStatusDatabase[keyof typeof HealthStatusDatabase];


export const HealthStatusDatabase = {
  connected: 'connected',
  unavailable: 'unavailable',
} as const;

export type HealthStatusRedis = typeof HealthStatusRedis[keyof typeof HealthStatusRedis];


export const HealthStatusRedis = {
  connected: 'connected',
  unavailable: 'unavailable',
  not_configured: 'not_configured',
} as const;

export interface HealthStatus {
  status: string;
  service: string;
  uptimeSeconds: number;
  telegram: HealthStatusTelegram;
  database: HealthStatusDatabase;
  redis: HealthStatusRedis;
}

export interface ConfidencePoint {
  label: string;
  value: number;
}

export interface DashboardSummary {
  monitoredChannels: number;
  activeChannels: number;
  detectedCalls: number;
  pendingReview: number;
  publishedCalls: number;
  duplicateQueue: number;
  winRate: number;
  uptimeSeconds: number;
  confidenceHistory: ConfidencePoint[];
}

export type ChannelKind = typeof ChannelKind[keyof typeof ChannelKind];


export const ChannelKind = {
  source: 'source',
  destination: 'destination',
} as const;

export type ChannelStatus = typeof ChannelStatus[keyof typeof ChannelStatus];


export const ChannelStatus = {
  monitoring: 'monitoring',
  paused: 'paused',
  error: 'error',
} as const;

export interface Channel {
  id: number;
  username: string;
  displayName: string;
  kind: ChannelKind;
  status: ChannelStatus;
  /** @nullable */
  lastSeenAt: string | null;
  messagesToday: number;
  /** @nullable */
  errorMessage?: string | null;
}

export interface ChannelInput {
  /**
     * @minLength 2
     * @maxLength 255
     */
  username: string;
  /** @maxLength 255 */
  displayName?: string;
}

export type ChannelUpdateStatus = typeof ChannelUpdateStatus[keyof typeof ChannelUpdateStatus];


export const ChannelUpdateStatus = {
  monitoring: 'monitoring',
  paused: 'paused',
  error: 'error',
} as const;

export interface ChannelUpdate {
  /** @maxLength 255 */
  displayName?: string;
  status?: ChannelUpdateStatus;
}

export type TokenCallStatus = typeof TokenCallStatus[keyof typeof TokenCallStatus];


export const TokenCallStatus = {
  pending: 'pending',
  published: 'published',
  rejected: 'rejected',
  duplicate: 'duplicate',
} as const;

export type TokenCallRisk = typeof TokenCallRisk[keyof typeof TokenCallRisk];


export const TokenCallRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

/**
 * @nullable
 */
export type TokenCallMilestone = typeof TokenCallMilestone[keyof typeof TokenCallMilestone] | null;


export const TokenCallMilestone = {
  '2x': '2x',
  '5x': '5x',
  '10x': '10x',
  '25x': '25x',
  '50x': '50x',
} as const;

export interface MarketSnapshot {
  /** @nullable */
  marketCap: number | null;
  /** @nullable */
  liquidity: number | null;
  /** @nullable */
  volume24h: number | null;
  /** @nullable */
  holders: number | null;
  /** @nullable */
  age: string | null;
  /** @nullable */
  priceChange24h?: number | null;
}

export interface TokenCall {
  id: number;
  ticker: string;
  tokenName: string;
  chain: string;
  contract: string;
  status: TokenCallStatus;
  confidence: number;
  risk: TokenCallRisk;
  narrative: string;
  observations: string[];
  sourceChannels: string[];
  detectedAt: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  milestone?: TokenCallMilestone;
  market: MarketSnapshot;
}

export type ReviewInputAction = typeof ReviewInputAction[keyof typeof ReviewInputAction];


export const ReviewInputAction = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface ReviewInput {
  action: ReviewInputAction;
}

export type ActivityEventKind = typeof ActivityEventKind[keyof typeof ActivityEventKind];


export const ActivityEventKind = {
  detected: 'detected',
  enriched: 'enriched',
  duplicate: 'duplicate',
  published: 'published',
  milestone: 'milestone',
  system: 'system',
  error: 'error',
} as const;

export type ActivityEventMetadata = { [key: string]: unknown };

export interface ActivityEvent {
  id: number;
  kind: ActivityEventKind;
  message: string;
  createdAt: string;
  metadata?: ActivityEventMetadata;
}

export type RuntimeSettingsLlmProvider = typeof RuntimeSettingsLlmProvider[keyof typeof RuntimeSettingsLlmProvider];


export const RuntimeSettingsLlmProvider = {
  openai: 'openai',
  anthropic: 'anthropic',
  grok: 'grok',
  gemini: 'gemini',
} as const;

export interface RuntimeSettings {
  destinationChannel: string;
  llmProvider: RuntimeSettingsLlmProvider;
  autoPublish: boolean;
  mediaRepost: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumConfidence: number;
  /**
     * @minimum 1
     * @maximum 720
     */
  duplicateWindowHours: number;
}

export type RuntimeSettingsUpdateLlmProvider = typeof RuntimeSettingsUpdateLlmProvider[keyof typeof RuntimeSettingsUpdateLlmProvider];


export const RuntimeSettingsUpdateLlmProvider = {
  openai: 'openai',
  anthropic: 'anthropic',
  grok: 'grok',
  gemini: 'gemini',
} as const;

export interface RuntimeSettingsUpdate {
  destinationChannel?: string;
  llmProvider?: RuntimeSettingsUpdateLlmProvider;
  autoPublish?: boolean;
  mediaRepost?: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumConfidence?: number;
  /**
     * @minimum 1
     * @maximum 720
     */
  duplicateWindowHours?: number;
}

export interface CredentialsStatus {
  telegramApiIdConfigured: boolean;
  telegramApiHashConfigured: boolean;
  telegramSessionConfigured: boolean;
  telegramBotTokenConfigured: boolean;
  geminiApiKeyConfigured: boolean;
}

export interface CredentialsUpdate {
  /** @minLength 1 */
  telegramApiId?: string;
  /** @minLength 1 */
  telegramApiHash?: string;
  /** @minLength 1 */
  telegramSession?: string;
  /** @minLength 1 */
  telegramBotToken?: string;
  /** @minLength 1 */
  geminiApiKey?: string;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
} as const;

export type LogEntryMetadata = { [key: string]: unknown };

export interface LogEntry {
  id: number;
  level: LogEntryLevel;
  service: string;
  message: string;
  createdAt: string;
  metadata?: LogEntryMetadata;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type LimitParameter = number;

export type GetDashboardActivityParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};

export type ListCallsParams = {
status?: ListCallsStatus;
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};

export type ListCallsStatus = typeof ListCallsStatus[keyof typeof ListCallsStatus];


export const ListCallsStatus = {
  all: 'all',
  pending: 'pending',
  published: 'published',
  rejected: 'rejected',
  duplicate: 'duplicate',
} as const;

export type ListLogsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};