# lib/api-zod/src/generated/types/channel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ARCC Signal Hub operational API
 * OpenAPI spec version: 0.1.0
 */
import type { ChannelKind } from './channelKind';
import type { ChannelStatus } from './channelStatus';

export interface Channel {
  id: number;
  username: string;
  displayName: string;
  kind: ChannelKind;
  status: ChannelStatus;
  /** @nullable */
  lastSeenAt: Date | null;
  messagesToday: number;
  /** @nullable */
  errorMessage?: string | null;
}