# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ARCC Signal Hub operational API
 * OpenAPI spec version: 0.1.0
 */
import type { ConfidencePoint } from './confidencePoint';

export interface DashboardSummary {
  monitoredChannels: number;
  activeChannels: number;
  detectedCalls: number;
  pendingReview: number;
  publishedCalls: number;
  duplicateQueue: number;
  winRate: number;
  uptimeSeconds: number;
  confidenceHistory: ConfidencePoint[];
}