# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ARCC Signal Hub operational API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusDatabase } from './healthStatusDatabase';
import type { HealthStatusRedis } from './healthStatusRedis';
import type { HealthStatusTelegram } from './healthStatusTelegram';

export interface HealthStatus {
  status: string;
  service: string;
  uptimeSeconds: number;
  telegram: HealthStatusTelegram;
  database: HealthStatusDatabase;
  redis: HealthStatusRedis;
}